Corrección de error en Delete

+++ Controllers/albumControllers/updateAndDelete.js
@@ -44,7 +44,13 @@ function deleteAlbum(req, res) {
                     Song.find({album: req.params.id})
                     .then((songs) => {
                         songs.forEach((delsong) => {
-                            fs.unlinkSync('Songs/' + delsong.file)
+                            if (delsong.file) {
+                                const path = path.join(__dirname, '../../Songs/', delsong.file);
+                                const exists = fs.existsSync(path);
+                                if (exists) {
+                                    fs.unlinkSync('Songs/' + delsong.file)
+                                }
+                            }
                         })
                     })
                     Song.deleteMany({ album: req.params.id })
